chore: 이미지 코드 구조 분리
